requirements.txt
streamlit
tensorflow
pillow
numpy
scikit-learn
matplotlib